Rename GUI plugin constant files
