+++ src/constants/plugins/buyPluginList.json
