+++ src/modules/UI/components/ControlPanel/Component/Main.js
@@ -16,7 +16,7 @@ import sweepIcon from '../../../../../assets/images/sidenav/sweep.png'
 import termsIcon from '../../../../../assets/images/sidenav/terms.png'
 import walletIcon from '../../../../../assets/images/sidenav/wallets.png'
 import * as Constants from '../../../../../constants/indexConstants.js'
-import { pluginUrlMap } from '../../../../../constants/plugins/buySellPlugins'
+import { pluginUrlMap } from '../../../../../constants/plugins/GuiPlugins.js'
 import s from '../../../../../locales/strings.js'
 import { scale } from '../../../../../util/scaling.js'
 import styles from '../style'
